(refactor): break oauth<->sdks cycle by layering token refresh into sdks
getValidAccessToken (the bearer-refresh primitive that sdkForConsole/
sdkForProject depend on) moves from auth/oauth.ts into sdks.ts where it
belongs, and the shared OAUTH2_CLIENT_ID/OAUTH2_SCOPES move to constants.ts.
sdks.ts no longer imports auth/oauth.ts, so auth/oauth.ts is free to build its
OAuth2 clients via services.ts getOauth2Service without a circular import.

Net dependency flow is now acyclic: constants/config (leaves) -> sdks ->
services -> oauth -> session/login. getValidAccessToken constructs its refresh
client directly (it sits below the services factory); revoke and the device
flow use getOauth2Service.
