+++ templates/cli/lib/constants.ts.twig
@@ -27,6 +27,10 @@ export const GITHUB_RELEASES_URL = `https://github.com/${GITHUB_REPO}/releases`;
 // API
 export const DEFAULT_ENDPOINT = '{{ spec.endpoint }}';
 
+// OAuth2
+export const OAUTH2_CLIENT_ID = "appwrite-cli";
+export const OAUTH2_SCOPES = "openid email profile";
+
 // Config resources
 export const CONFIG_RESOURCE_KEYS = [
   "databases",
